# openapi/fees/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    name: Kita Knirpsenstadt
  description: API für Beitragsverwaltung der Kita Knirpsenstadt
  title: Kita Knirpsenstadt Fees API
  version: 1.0.0
paths:
...
  /import/confirm:
    post:
      description: Confirm multiple transaction-fee matches from suggestions
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmMatchRequest"
        description: Matches to confirm
        required: true
      responses:
        "200":
          description: Confirmation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmMatchResponse"
        "400":
          description: Invalid request body or IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Confirm suggested matches
      tags:
        - Import
  /import/history:
    get:
      description: Get a paginated list of past CSV imports
      parameters:
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
            default: 1
        - description: Items per page
          in: query
          name: perPage
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Import history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportHistoryList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get import history
      tags:
        - Import
...
  /import/rescan:
    post:
      description: Re-run matching algorithm on all unmatched transactions.
        High-confidence matches (95%+) are automatically confirmed.
      responses:
        "200":
          description: Rescan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RescanResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Rescan unmatched transactions
      tags:
        - Import
...
  /import/upload:
    post:
      description: Upload and process a bank statement CSV file to match transactions
        with fees
      requestBody:
        $ref: "#/components/requestBodies/postChildrenImportParse"
      responses:
        "200":
          description: Upload and processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/service.ImportResult"
        "400":
          description: No file provided or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File larger than 5MB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Upload a CSV file for import
      tags:
        - Import
...
servers:
  - url: http://localhost:8081/api/fees/v1
  - url: https://localhost:8081/api/fees/v1
components:
  requestBodies:
...
    postChildrenImportParse:
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              file:
                description: CSV file (max 5MB)
                type: string
                format: binary
            required:
              - file
      required: true
  securitySchemes:
    BearerAuth:
      description: JWT Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    ConfirmMatchRequest:
      description: Request body for confirming transaction-fee matches
      properties:
        matches:
          items:
            $ref: "#/components/schemas/MatchConfirmation"
          type: array
      required:
        - matches
      type: object
    ConfirmMatchResponse:
      description: Result of confirming matches
      properties:
        confirmed:
          example: 5
          type: integer
        failed:
          example: 0
          type: integer
      required:
        - confirmed
        - failed
      type: object
...
    ErrorResponse:
      description: Standard error response format
      properties:
        code:
          example: 400
          type: integer
        error:
          example: Bad Request
          type: string
        message:
          example: invalid request body
          type: string
      required:
        - code
        - error
        - message
      type: object
...
    ImportError:
      properties:
        amount:
          type: number
        bookingDate:
          type: string
        message:
          type: string
        payerName:
          type: string
      required:
        - amount
        - message
      type: object
    ImportHistoryList:
      description: Paginated import history
      properties:
        data:
          items:
            $ref: "#/components/schemas/domain.ImportBatch"
          type: array
        page:
          example: 1
          type: integer
        perPage:
          example: 20
          type: integer
        total:
          example: 10
          type: integer
        totalPages:
          example: 1
          type: integer
      required:
        - data
        - page
        - perPage
        - total
        - totalPages
      type: object
...
    MatchConfirmation:
      description: Single match confirmation
      properties:
        expectationId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        transactionId:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
      required:
        - expectationId
        - transactionId
      type: object
    MatchSuggestion:
      description: Suggested match for manual review
      properties:
        confidence:
          example: 0.85
          type: number
        expectationId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        expectationInfo:
          example: Betreuungsgebühr März 2024 - Max Mustermann
          type: string
        reason:
          example: Name match
          type: string
        transactionId:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        transactionInfo:
          example: SEPA-Überweisung Max Mustermann
          type: string
      required:
        - confidence
        - expectationId
        - expectationInfo
        - reason
        - transactionId
        - transactionInfo
      type: object
...
    RescanResponse:
      description: Rescan result with new match suggestions
      properties:
        autoMatched:
          example: 150
          type: integer
        errors:
          description: Errors lists transactions whose warning or automatic match could
            not be saved.
          items:
            $ref: "#/components/schemas/ImportError"
          type: array
        newMatches:
          example: 5
          type: integer
        scanned:
          example: 226
          type: integer
        suggestions:
          items:
            $ref: "#/components/schemas/MatchSuggestion"
          type: array
      required:
        - autoMatched
        - errors
        - newMatches
        - scanned
      type: object
...
    domain.BankTransaction:
      properties:
        amount:
          type: number
        bookingDate:
          type: string
        currency:
          type: string
        description:
          type: string
        hiddenAt:
          type: string
        hiddenBy:
          type: string
        id:
          type: string
        importBatchId:
          type: string
        importedAt:
          type: string
        isHidden:
          type: boolean
        matchedAmount:
          description: MatchedAmount is the total amount of this transaction already
            allocated to fees.
          type: number
        matches:
          description: Joined fields
          items:
            $ref: "#/components/schemas/domain.PaymentMatch"
          type: array
        payerIban:
          type: string
        payerName:
          type: string
        valueDate:
          type: string
      required:
        - amount
        - bookingDate
        - currency
        - id
        - importedAt
        - isHidden
        - valueDate
      type: object
    domain.Child:
      properties:
        birthDate:
          type: string
        careHours:
          type: integer
        city:
          type: string
        createdAt:
          type: string
        entryDate:
          type: string
        exitDate:
          type: string
        firstName:
          type: string
        household:
          $ref: "#/components/schemas/domain.Household"
        householdId:
          type: string
        id:
          type: string
        isActive:
          type: boolean
        lastName:
          type: string
        legalHours:
          description: Derived from the history tables (see type comment), not stored on
            fees.children.
          type: integer
        legalHoursUntil:
          type: string
        memberNumber:
          type: string
        openFeesCount:
          description: Computed fields (not stored in DB, populated by queries)
          type: integer
        parents:
          description: Loaded relations (not stored in DB)
          items:
            $ref: "#/components/schemas/domain.Parent"
          type: array
        postalCode:
          type: string
        street:
          type: string
        streetNo:
          type: string
        updatedAt:
          type: string
      required:
        - birthDate
        - createdAt
        - entryDate
        - firstName
        - id
        - isActive
        - lastName
        - memberNumber
        - updatedAt
      type: object
...
    domain.FeeExpectation:
      properties:
        amount:
          type: number
        child:
          allOf:
            - $ref: "#/components/schemas/domain.Child"
          description: Joined fields
        childId:
          type: string
        createdAt:
          type: string
        dueDate:
          type: string
        feeType:
          $ref: "#/components/schemas/domain.FeeType"
        householdId:
          type: string
        id:
          type: string
        isPaid:
          type: boolean
        matchedAmount:
          description: Total matched across all transactions
          type: number
        matchedBy:
          $ref: "#/components/schemas/domain.PaymentMatch"
        month:
          description: nil for yearly fees
          type: integer
        paidAt:
          type: string
        partialMatches:
          description: All transactions covering this fee
          items:
            $ref: "#/components/schemas/domain.PaymentMatch"
          type: array
        reconciliationYear:
          description: "For Kalendarjahresabrechnung: the year this Nachzahlung is for"
          type: integer
        remaining:
          description: Amount still needed
          type: number
        reminderForId:
          description: "For REMINDER type: links to the original fee"
          type: string
        year:
          type: integer
      required:
        - amount
        - childId
        - createdAt
        - dueDate
        - feeType
        - id
        - isPaid
        - year
      type: object
...
    domain.FeeType:
      enum:
        - MEMBERSHIP
        - FOOD
        - CHILDCARE
        - REMINDER
      type: string
      x-enum-comments:
        FeeTypeChildcare: Platzgeld (monthly, variable, for U3)
        FeeTypeFood: Essensgeld (monthly, 45.40 EUR)
        FeeTypeMembership: Vereinsbeitrag (yearly, 30 EUR)
        FeeTypeReminder: Mahngebühr (10 EUR, manually triggered)
      x-enum-descriptions:
        - Vereinsbeitrag (yearly, 30 EUR)
        - Essensgeld (monthly, 45.40 EUR)
        - Platzgeld (monthly, variable, for U3)
        - Mahngebühr (10 EUR, manually triggered)
      x-enum-varnames:
        - FeeTypeMembership
        - FeeTypeFood
        - FeeTypeChildcare
        - FeeTypeReminder
    domain.Household:
      properties:
        annualHouseholdIncome:
          type: number
        children:
          items:
            $ref: "#/components/schemas/domain.Child"
          type: array
        childrenCountForFees:
          description: Override for sibling discount calculation
          type: integer
        createdAt:
          type: string
        id:
          type: string
        incomeStatus:
          $ref: "#/components/schemas/domain.IncomeStatus"
        membershipAssignmentStatus:
          $ref: "#/components/schemas/domain.MembershipAssignmentStatus"
        membershipParentId:
          type: string
        name:
          description: e.g. "Familie Müller" - auto-generated or manual
          type: string
        parents:
          description: Loaded relations (not stored in DB)
          items:
            $ref: "#/components/schemas/domain.Parent"
          type: array
        updatedAt:
          type: string
      required:
        - createdAt
        - id
        - incomeStatus
        - membershipAssignmentStatus
        - name
        - updatedAt
      type: object
...
    domain.ImportBatch:
      properties:
        dateFrom:
          type: string
        dateTo:
          type: string
        errorCount:
          description: |-
            ErrorCount is the number of rows that failed; Errors holds at most
            MaxStoredImportErrors of them.
          type: integer
        errors:
          items:
            $ref: "#/components/schemas/ImportError"
          type: array
        fileName:
          type: string
        id:
          type: string
        importedAt:
          type: string
        importedBy:
          type: string
        importedByEmail:
          type: string
        matchedCount:
          type: integer
        transactionCount:
          type: integer
      required:
        - errorCount
        - errors
        - fileName
        - id
        - importedAt
        - importedBy
        - importedByEmail
        - matchedCount
        - transactionCount
      type: object
...
    domain.IncomeStatus:
      enum:
        - ""
        - PROVIDED
        - MAX_ACCEPTED
        - PENDING
        - NOT_REQUIRED
        - HISTORIC
        - FOSTER_FAMILY
      type: string
      x-enum-varnames:
        - IncomeStatusUnknown
        - IncomeStatusProvided
        - IncomeStatusMaxAccepted
        - IncomeStatusPending
        - IncomeStatusNotRequired
        - IncomeStatusHistoric
        - IncomeStatusFosterFamily
...
    domain.MatchSuggestion:
      properties:
        child:
          $ref: "#/components/schemas/domain.Child"
        confidence:
          type: number
        detectedType:
          $ref: "#/components/schemas/domain.FeeType"
        expectation:
          $ref: "#/components/schemas/domain.FeeExpectation"
        expectations:
          description: For combined matches (e.g., fee + reminder)
          items:
            $ref: "#/components/schemas/domain.FeeExpectation"
          type: array
        matchedBy:
          description: '"member_number", "name", "amount", "combined"'
          type: string
        transaction:
          $ref: "#/components/schemas/domain.BankTransaction"
      required:
        - confidence
        - matchedBy
        - transaction
      type: object
    domain.MatchType:
      enum:
        - AUTO
        - MANUAL
      type: string
      x-enum-varnames:
        - MatchTypeAuto
        - MatchTypeManual
    domain.Member:
      properties:
        city:
          type: string
        createdAt:
          type: string
        email:
          type: string
        firstName:
          type: string
        household:
          allOf:
            - $ref: "#/components/schemas/domain.Household"
          description: Loaded relations (not stored in DB)
        householdId:
          type: string
        id:
          type: string
        isActive:
          type: boolean
        lastName:
          type: string
        memberNumber:
          type: string
        membershipEnd:
          type: string
        membershipStart:
          type: string
        phone:
          type: string
        postalCode:
          type: string
        street:
          type: string
        streetNo:
          type: string
        updatedAt:
          type: string
      required:
        - createdAt
        - firstName
        - id
        - isActive
        - lastName
        - memberNumber
        - membershipStart
        - updatedAt
      type: object
    domain.MembershipAssignmentStatus:
      enum:
        - ASSUMED
        - CONFIRMED
      type: string
      x-enum-varnames:
        - MembershipAssignmentStatusAssumed
        - MembershipAssignmentStatusConfirmed
...
    domain.Parent:
      properties:
        annualHouseholdIncome:
          description: "DEPRECATED: Income fields moved to Household. Kept for backwards
            compatibility during migration."
          type: number
        birthDate:
          type: string
        children:
          description: Loaded relations (not stored in DB)
          items:
            $ref: "#/components/schemas/domain.Child"
          type: array
        city:
          type: string
        createdAt:
          type: string
        email:
          type: string
        firstName:
          type: string
        household:
          $ref: "#/components/schemas/domain.Household"
        householdId:
          type: string
        id:
          type: string
        incomeStatus:
          $ref: "#/components/schemas/domain.IncomeStatus"
        lastName:
          type: string
        member:
          $ref: "#/components/schemas/domain.Member"
        memberId:
          description: Link to Member if this parent is also a member
          type: string
        phone:
          type: string
        postalCode:
          type: string
        street:
          type: string
        streetNo:
          type: string
        updatedAt:
          type: string
      required:
        - createdAt
        - firstName
        - id
        - incomeStatus
        - lastName
        - updatedAt
      type: object
    domain.PaymentMatch:
      properties:
        amount:
          type: number
        confidence:
          type: number
        expectation:
          $ref: "#/components/schemas/domain.FeeExpectation"
        expectationId:
          type: string
        id:
          type: string
        matchType:
          $ref: "#/components/schemas/domain.MatchType"
        matchedAt:
          type: string
        matchedBy:
          type: string
        transaction:
          allOf:
            - $ref: "#/components/schemas/domain.BankTransaction"
          description: Joined fields
        transactionId:
          type: string
      required:
        - amount
        - expectationId
        - id
        - matchType
        - matchedAt
        - transactionId
      type: object
    domain.ResolutionType:
      enum:
        - dismissed
        - matched
        - auto_resolved
      type: string
      x-enum-comments:
        ResolutionTypeAutoResolved: Automatically resolved
        ResolutionTypeDismissed: Manually dismissed with a reason
        ResolutionTypeMatched: Resolved by creating a match
      x-enum-descriptions:
        - Manually dismissed with a reason
        - Resolved by creating a match
        - Automatically resolved
      x-enum-varnames:
        - ResolutionTypeDismissed
        - ResolutionTypeMatched
        - ResolutionTypeAutoResolved
    domain.TransactionWarning:
      properties:
        actualAmount:
          type: number
        child:
          $ref: "#/components/schemas/domain.Child"
        childId:
          type: string
        createdAt:
          type: string
        expectedAmount:
          type: number
        id:
          type: string
        matchedFee:
          allOf:
            - $ref: "#/components/schemas/domain.FeeExpectation"
          description: "For LATE_PAYMENT: the fee that was matched"
        matchedFeeId:
          description: "For LATE_PAYMENT: the fee that was matched"
          type: string
        message:
          type: string
        resolutionNote:
          type: string
        resolutionType:
          $ref: "#/components/schemas/domain.ResolutionType"
        resolvedAt:
          type: string
        resolvedBy:
          type: string
        transaction:
          allOf:
            - $ref: "#/components/schemas/domain.BankTransaction"
          description: Joined fields
        transactionId:
          type: string
        warningType:
          $ref: "#/components/schemas/domain.WarningType"
      required:
        - createdAt
        - id
        - message
        - transactionId
        - warningType
      type: object
    domain.WarningType:
      enum:
        - NO_MATCHING_FEE
        - UNEXPECTED_AMOUNT
        - PARTIAL_PAYMENT
        - OVERPAYMENT
        - POSSIBLE_BULK
        - DUPLICATE_PAYMENT
        - LATE_PAYMENT
        - MULTIPLE_OPEN_FEES
      type: string
      x-enum-comments:
        WarningTypeDuplicatePayment: Fee already paid, this might be duplicate
        WarningTypeLatePayment: Payment received after the 15th of fee month
        WarningTypeMultipleOpenFees: Multiple unpaid fees exist, manual review needed
        WarningTypeNoMatchingFee: Trusted IBAN but no open fee found
        WarningTypeOverpayment: Amount is more than expected
        WarningTypePartialPayment: Amount is less than expected
        WarningTypePossibleBulk: Amount could be multiple fees combined
        WarningTypeUnexpectedAmount: Amount doesn't match any expected fee
      x-enum-descriptions:
        - Trusted IBAN but no open fee found
        - Amount doesn't match any expected fee
        - Amount is less than expected
        - Amount is more than expected
        - Amount could be multiple fees combined
        - Fee already paid, this might be duplicate
        - Payment received after the 15th of fee month
        - Multiple unpaid fees exist, manual review needed
      x-enum-varnames:
        - WarningTypeNoMatchingFee
        - WarningTypeUnexpectedAmount
        - WarningTypePartialPayment
        - WarningTypeOverpayment
        - WarningTypePossibleBulk
        - WarningTypeDuplicatePayment
        - WarningTypeLatePayment
        - WarningTypeMultipleOpenFees
...
    service.ImportResult:
      properties:
        autoMatched:
          type: integer
        batchId:
          type: string
        blacklisted:
          type: integer
        errors:
          description: >-
            Errors lists rows that could not be saved or processed; they are
            also

            stored on the import batch.
          items:
            $ref: "#/components/schemas/ImportError"
          type: array
        fileName:
          type: string
        imported:
          type: integer
        skipped:
          type: integer
        suggestions:
          items:
            $ref: "#/components/schemas/domain.MatchSuggestion"
          type: array
        totalRows:
          type: integer
        warningList:
          items:
            $ref: "#/components/schemas/domain.TransactionWarning"
          type: array
        warnings:
          type: integer
      required:
        - autoMatched
        - batchId
        - blacklisted
        - errors
        - fileName
        - imported
        - skipped
        - suggestions
        - totalRows
        - warnings
      type: object
...